# specs/001-weatherstation-backend/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: WeatherStation Backend API (Red de Estaciones)
  description: >
    API REST del backend central de la Red de Estaciones Meteorológicas
    Inteligentes (CLIMBOT). Único componente con acceso a la base de datos
    PostgreSQL (almacén único: sistema + lecturas meteorológicas) y a Gemini (IA).
    Las estaciones ESP32 publican sus lecturas a través de `/api/device/**` (nunca
    a un almacén externo). Autenticación JWT (Bearer) con roles ADMIN, RESPONSABLE,
    INVESTIGADOR y USUARIO para usuarios, y token de estación → JWT de dispositivo
    para los ESP32.
  version: 3.1.0
  contact:
    name: WeatherStation
servers:
  - url: http://localhost:8080
    description: Desarrollo local
  - url: https://weatherstation-backend.onrender.com
    description: Producción (Render)
...
security:
  - bearerAuth: []
paths:
...
  /api/public/weather/latest:
    get:
      tags: [Público]
      summary: Última lectura por estación (para mapa/listado público)
      security: []
      parameters:
        - { in: query, name: municipio, schema: { type: string } }
      responses:
        '200':
          description: Clima actual público
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PublicEstacionResponse' }
  /api/public/statistics:
    get:
      tags: [Público]
      summary: Estadísticas agregadas no sensibles (red / municipio / escuela)
      security: []
      parameters:
        - { in: query, name: municipio, schema: { type: string } }
        - { in: query, name: escuelaId, schema: { type: string, format: uuid } }
        - { in: query, name: desde, schema: { type: string, format: date-time } }
        - { in: query, name: hasta, schema: { type: string, format: date-time } }
      responses:
        '200':
          description: Estadísticas públicas
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PublicEstadisticaResponse' }

  # ──────────────────────────── ESCUELAS ─────────────────────────────
...
  /estaciones:
    get:
      tags: [Estaciones]
      summary: Listar estaciones (visibilidad según rol)
      description: >
        USUARIO/INVESTIGADOR ven solo APPROVED+habilitadas; RESPONSABLE además las
        de su escuela en cualquier estado; ADMIN todas. Filtros opcionales.
      parameters:
        - { in: query, name: escuelaId, schema: { type: string, format: uuid } }
        - { in: query, name: estado, schema: { $ref: '#/components/schemas/EstadoEstacion' } }
      responses:
        '200':
          description: Lista de estaciones
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EstacionResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Estaciones]
      summary: Registrar una estación (ADMIN o RESPONSABLE de la escuela) → PENDING
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EstacionRegistroRequest' }
      responses:
        '201':
          description: Estación creada en estado PENDING
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EstacionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /estaciones/{id}:
    parameters: [ { $ref: '#/components/parameters/EstacionId' } ]
    get:
      tags: [Estaciones]
      summary: Detalle de una estación
      responses:
        '200':
          description: Estación
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EstacionResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Estaciones]
      summary: Actualizar metadata de la estación (ADMIN o RESPONSABLE de su escuela)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EstacionUpdateRequest' }
      responses:
        '200':
          description: Estación actualizada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EstacionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Estaciones]
      summary: Eliminar una estación (ADMIN)
      responses:
        '204': { description: Eliminada }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /estaciones/{id}/actual:
    parameters: [ { $ref: '#/components/parameters/EstacionId' } ]
    get:
      tags: [Estaciones]
      summary: Última lectura de la estación (cualquier rol)
      responses:
        '200':
          description: Lectura actual
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LecturaResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
  /estaciones/{id}/historial:
    parameters:
      - { $ref: '#/components/parameters/EstacionId' }
      - in: query
        name: desde
        schema: { type: string, format: date-time }
        description: Inicio del rango (ISO-8601). Por defecto, 24 h atrás.
      - in: query
        name: hasta
        schema: { type: string, format: date-time }
        description: Fin del rango (ISO-8601). Por defecto, ahora.
    get:
      tags: [Estaciones]
      summary: Historial por rango (RESPONSABLE de su escuela, INVESTIGADOR, ADMIN)
      responses:
        '200':
          description: Lecturas ordenadas cronológicamente
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LecturaResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /estaciones/{id}/ultimas24h:
    parameters: [ { $ref: '#/components/parameters/EstacionId' } ]
    get:
      tags: [Estaciones]
      summary: Lecturas de las últimas 24 horas (RESPONSABLE, INVESTIGADOR, ADMIN)
      responses:
        '200':
          description: Lecturas de las últimas 24 h
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/LecturaResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /estaciones/{id}/estadisticas:
    parameters:
      - { $ref: '#/components/parameters/EstacionId' }
      - { in: query, name: desde, schema: { type: string, format: date-time } }
      - { in: query, name: hasta, schema: { type: string, format: date-time } }
    get:
      tags: [Estaciones]
      summary: Estadísticas agregadas del rango (RESPONSABLE, INVESTIGADOR, ADMIN)
      responses:
        '200':
          description: Estadísticas
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EstadisticasResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /solicitudes:
    get:
      tags: [Solicitudes]
      summary: Listar solicitudes de alta (ADMIN; RESPONSABLE solo las de su escuela)
      parameters:
        - { in: query, name: estado, schema: { $ref: '#/components/schemas/EstadoSolicitud' } }
        - { in: query, name: escuelaId, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Lista de solicitudes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/SolicitudResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /solicitudes/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [Solicitudes]
      summary: Ver una solicitud (ADMIN o RESPONSABLE de su escuela)
      responses:
        '200':
          description: Solicitud
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SolicitudResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /solicitudes/{id}/aprobar:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [Solicitudes]
      summary: Aprobar la solicitud (ADMIN) → crea estación APPROVED + token (una vez)
      responses:
        '200':
          description: Solicitud aprobada; token de la nueva estación (una vez)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/StationTokenResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /solicitudes/{id}/rechazar:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [Solicitudes]
      summary: Rechazar la solicitud (ADMIN)
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: '#/components/schemas/EstacionAccionRequest' }
      responses:
        '200':
          description: Solicitud rechazada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SolicitudResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ──────────────────────────── USUARIOS ─────────────────────────────
...
  /ia/preguntar:
    post:
      tags: [IA]
      summary: Pregunta en lenguaje natural sobre una estación (cualquier rol)
      description: Rate limited. Fundamentado en datos reales de PostgreSQL.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IaPreguntaRequest' }
      responses:
        '200':
          description: Respuesta de IA
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IaResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /ia/resumen:
    post:
      tags: [IA]
      summary: Resumen del clima de una ventana temporal (cualquier rol)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IaResumenRequest' }
      responses:
        '200':
          description: Resumen
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IaResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /ia/prediccion:
    post:
      tags: [IA]
      summary: Estimación cualitativa basada en tendencias (cualquier rol)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/IaPrediccionRequest' }
      responses:
        '200':
          description: Predicción cualitativa
          content:
            application/json:
              schema: { $ref: '#/components/schemas/IaResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }

  # ──────────────────────────── ALERTAS ──────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT de usuario (claim `rol`).
    deviceAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT de dispositivo emitido por /api/device/auth (claim `type=DEVICE`).
  parameters:
    EstacionId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
      description: UUID público de la estación
      example: 7b1f2c34-5d6e-4a8b-9c0d-1e2f3a4b5c6d
  responses:
    BadRequest:
      description: Datos de entrada inválidos
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: No autenticado o credenciales inválidas
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Forbidden:
      description: Rol/recurso no autorizado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    TooManyRequests:
      description: Límite de peticiones excedido (IA o dispositivo)
      headers:
        Retry-After:
          schema: { type: integer }
          description: Segundos hasta poder reintentar
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    ServiceUnavailable:
      description: Servicio externo (Gemini) o base de datos no disponible
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
...
    EstadoEstacion:
      type: string
      enum: [PENDING, APPROVED, REJECTED, DISABLED, MAINTENANCE]
    Conectividad:
      type: string
      enum: [ONLINE, OFFLINE]
      description: Derivado de ultimaConexion + umbral; no es un estado almacenado.
    EstadoSolicitud:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
...
    EstacionRegistroRequest:
      type: object
      required: [nombre, escuelaId]
      properties:
        nombre: { type: string, minLength: 2, maxLength: 120 }
        escuelaId: { type: string, format: uuid }
        ubicacion: { type: string }
        municipio: { type: string }
        latitud: { type: number, format: double }
        longitud: { type: number, format: double }
        firmware: { type: string, maxLength: 40 }
    EstacionUpdateRequest:
      type: object
      properties:
        nombre: { type: string, maxLength: 120 }
        ubicacion: { type: string }
        municipio: { type: string }
        latitud: { type: number, format: double }
        longitud: { type: number, format: double }
        responsableId: { type: string, format: uuid }
    EstacionAccionRequest:
      type: object
      properties:
        motivo: { type: string, maxLength: 300 }
    EstacionResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        uuid: { type: string, format: uuid }
        nombre: { type: string }
        descripcion: { type: string }
        escuelaId: { type: string, format: uuid }
        escuelaNombre: { type: string }
        responsableId: { oneOf: [ { type: string, format: uuid }, { type: 'null' } ] }
        responsableNombre: { oneOf: [ { type: string }, { type: 'null' } ] }
        ubicacion: { type: string }
        municipio: { type: string }
        latitud: { type: number, format: double }
        longitud: { type: number, format: double }
        altitud: { type: number, format: double }
        firmware: { type: string }
        hardware: { type: string }
        ultimoRssi: { oneOf: [ { type: integer }, { type: 'null' } ] }
        estado: { $ref: '#/components/schemas/EstadoEstacion' }
        conectividad: { $ref: '#/components/schemas/Conectividad' }
        fechaRegistro: { type: string, format: date-time }
        ultimaConexion: { oneOf: [ { type: string, format: date-time }, { type: 'null' } ] }
        enLinea: { type: boolean }
        ultimaLectura:
          oneOf:
            - { $ref: '#/components/schemas/LecturaResponse' }
            - { type: 'null' }
    StationTokenResponse:
      type: object
      description: El token en claro se muestra una única vez y no se vuelve a exponer.
      properties:
        estacionId: { type: string, format: uuid }
        uuid: { type: string, format: uuid }
        token: { type: string, description: Token permanente de la estación (en claro, solo aquí) }
        generadoEn: { type: string, format: date-time }
        aviso: { type: string, example: "Guarde este token: no se volverá a mostrar." }
...
    SolicitudResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        uuidPropuesto: { type: string, format: uuid }
        nombre: { type: string }
        escuelaId: { type: string, format: uuid }
        escuelaNombre: { type: string }
        ubicacion: { type: string }
        municipio: { type: string }
        firmware: { type: string }
        estado: { $ref: '#/components/schemas/EstadoSolicitud' }
        motivoRechazo: { oneOf: [ { type: string }, { type: 'null' } ] }
        estacionId: { oneOf: [ { type: string, format: uuid }, { type: 'null' } ] }
        createdAt: { type: string, format: date-time }
        resueltaEn: { oneOf: [ { type: string, format: date-time }, { type: 'null' } ] }
...
    PublicEstacionResponse:
      type: object
      properties:
        uuid: { type: string, format: uuid }
        nombre: { type: string }
        municipio: { type: string }
        latitud: { type: number, format: double }
        longitud: { type: number, format: double }
        escuelaNombre: { type: string }
        conectividad: { $ref: '#/components/schemas/Conectividad' }
        ultimaLectura:
          oneOf:
            - { $ref: '#/components/schemas/LecturaResponse' }
            - { type: 'null' }
    PublicEstadisticaResponse:
      type: object
      properties:
        ambito: { type: string, example: "MUNICIPIO:Chilpancingo" }
        desde: { type: string, format: date-time }
        hasta: { type: string, format: date-time }
        estaciones: { type: integer }
        muestras: { type: integer }
        temperatura: { $ref: '#/components/schemas/Agregado' }
        humedad: { $ref: '#/components/schemas/Agregado' }
        presion: { $ref: '#/components/schemas/Agregado' }
        vientoKmh: { $ref: '#/components/schemas/Agregado' }
        lluviaTotalMm: { type: number, format: double }
...
    LecturaResponse:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        temperatura: { type: number, format: double, example: 24.5 }
        humedad: { type: number, format: double, example: 60 }
        presion: { type: number, format: double, example: 1013.2 }
        vientoKmh: { type: number, format: double, example: 12.3 }
        vientoDir: { type: string, example: NE }
        vientoGrados:
          oneOf:
            - { type: number, format: double }
            - { type: 'null' }
          description: Solo presente en la lectura actual; null en historial
        lluviaMm: { type: number, format: double, example: 0.0 }
    Agregado:
      type: object
      properties:
        min: { type: number, format: double }
        max: { type: number, format: double }
        promedio: { type: number, format: double }
    EstadisticasResponse:
      type: object
      properties:
        estacionId: { type: string, format: uuid }
        desde: { type: string, format: date-time }
        hasta: { type: string, format: date-time }
        muestras: { type: integer }
        temperatura: { $ref: '#/components/schemas/Agregado' }
        humedad: { $ref: '#/components/schemas/Agregado' }
        presion: { $ref: '#/components/schemas/Agregado' }
        vientoKmh: { $ref: '#/components/schemas/Agregado' }
        lluviaTotalMm: { type: number, format: double }
    IaPreguntaRequest:
      type: object
      required: [estacionId, pregunta]
      properties:
        estacionId: { type: string, format: uuid }
        pregunta: { type: string, maxLength: 500, example: "¿Va a llover hoy?" }
    IaResumenRequest:
      type: object
      required: [estacionId]
      properties:
        estacionId: { type: string, format: uuid }
        horas: { type: integer, minimum: 1, maximum: 168, default: 24 }
    IaPrediccionRequest:
      type: object
      required: [estacionId]
      properties:
        estacionId: { type: string, format: uuid }
    IaResponse:
      type: object
      properties:
        estacionId: { type: string, format: uuid }
        respuesta: { type: string }
        rangoDesde: { type: string, format: date-time }
        rangoHasta: { type: string, format: date-time }
        muestrasUsadas: { type: integer }
        advertencias:
          type: array
          items: { type: string }
        generadoEn: { type: string, format: date-time }
...
    ApiError:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        code: { type: string, example: ESTACION_NO_APROBADA }
        message: { type: string }
        path: { type: string }
        fieldErrors:
          type: object
          additionalProperties: { type: string }